requirements.txt
# ===== Backend =====
fastapi==0.110.0
uvicorn==0.27.1

# ===== Core AI / RAG =====
langchain==0.1.20
langchain-community==0.0.38

# ===== Vector DB =====
chromadb==0.4.24

#====Embedding====
sentence-transformers==2.7.0
# ===== LLM =====
groq==0.5.0
# ===== Core deps =====
pydantic==1.10.13
numpy==1.26.4
SQLAlchemy==2.0.29
aiohttp==3.9.5

# ===== Utils =====
requests==2.31.0
python-dotenv==1.0.1
tqdm==4.66.2

